requirement.txt
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
cycler==0.10.0
Cython==0.29.23
Flask==2.0.0
Flask-WTF==0.14.3
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.0
joblib==1.0.1
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==2.0.0
matplotlib==3.4.2
multitasking==0.0.9
mysql-connector-python==8.0.25
mysqlclient==2.0.3
nltk==3.6.2
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
polygon-api-client==0.1.11
praw==7.2.0
prawcore==2.0.0
protobuf==3.17.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.15
threadpoolctl==2.1.0
tqdm==4.60.0
update-checker==0.18.0
urllib3==1.26.4
websocket-client==0.59.0
websockets==9.0.2
Werkzeug==2.0.0
WTForms==2.3.3
yfinance==0.1.59
